backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Azure
azure-storage-blob==12.19.0
openai==1.3.0

# Document Processing
PyMuPDF==1.23.8
python-docx==1.1.0
pypdf2==3.0.1

# ML & Data
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
tiktoken==0.5.2

# Email
secure-smtplib==0.1.1

# PDF Generation
reportlab==4.0.7

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.2